% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GECO_score.R
\name{GECO_score}
\alias{GECO_score}
\title{Score predefined clusters using GECO.}
\usage{
GECO_score(clusters, gt_dir)
}
\arguments{
\item{clusters}{A data.frame with two columns: the first column is the name
of a gene and the second is the cluster label/name for that gene.}

\item{gt_dir}{The absolute path to the directory containing the ground truth
gene sets.}
}
\value{
A table with two columns: the first column is the name of the ground
truth set and the second is the GECO score for that ground truth set.
}
\description{
This function takes a set of user-defined clusters and a set of ground truth
genes and scores them using the GECO metric. The gene names within the
clusters and the gene names within the ground truth gene sets must match -
identical naming conventions must be used for both.
}
\examples{
# Load clusters
ex_clusters <- clusters_df
# Define ground truth gene set directory
GT_dir <- '/path/to/gt_gene_sets'
# Score the clusters
clusters <- GECO_score(ex_clusters, GT_dir)
}
