% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_clusters.R
\name{generate_clusters}
\alias{generate_clusters}
\title{Create the clusters.}
\usage{
generate_clusters(df, kmin, kmax, ktot, num_iter, km_algo)
}
\arguments{
\item{df}{A dataframe containing the normalized reads}

\item{kmin}{An integer indicating the minimum number of clusters to generate.
By default, this is set to 10.}

\item{kmax}{An integer indicating the maximum number of clusters to generate.
By default, this is set to 150.}

\item{ktot}{An integer indicating how many unique k-values to generate.
By default, this is set to 15. This produces 15 values ranging from kmin up
to kmax. Increasing this number will significantly impact performance.}

\item{num_iter}{An integer indicating the number or cluster iterations to
generate. By default, this is set to 10. This will perform the same k-means
clustering multiple times to account for the stochastic nature of the k-means
algorithm, resulting in a mean quality value in the final step that is more
reliable than a single iteration would be. Lowering this value will
negatively affect the GECO quality assessment, raising it will impact
performance.}

\item{km_algo}{A string indicating which k-means algorithm to use. By
default, this is set to 'Hartigan-Wong'.}
}
\value{
A list containing each iteration of the clustering performed. Within
each of the iterations are the kmeans objects for use in the second step e.g.
score_clusters(clusters).
}
\description{
This function takes the normalized data (TPM/FPKM & feature scaled) and uses
the k-means function to generate an iterative series of clusters to identify
a potentially optimal number of clusters for the dataset.
}
