% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yChromSet}
\alias{yChromSet}
\title{Protein coding genes found on the Y-chromosome. These genes are significantly
expressed in the 'rnaSeq' dataset provided. They are to be used as the sample
ground truth set in exploring the GECO package.}
\format{
An object of class \code{character} of length 9.
}
\usage{
yChromSet
}
\description{
Protein coding genes found on the Y-chromosome. These genes are significantly
expressed in the 'rnaSeq' dataset provided. They are to be used as the sample
ground truth set in exploring the GECO package.
}
\keyword{datasets}
