% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_quality.R
\name{assess_quality}
\alias{assess_quality}
\title{Generate the cluster quality figure.}
\usage{
assess_quality(GECO_scores)
}
\arguments{
\item{GECO_scores}{The scores dataframe output by score_clusters.}
}
\value{
A ggplot2 object ready for plotting e.g. plot(returned_object)
}
\description{
This function takes the GECO scores table generated by the score_clusters
function and finds the ROC/AUC values for the scores. These scores are then
arranged into a dataframe for display with ggplot2.
}
