% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_clusters.R
\name{score_clusters}
\alias{score_clusters}
\title{Score the k-means clusters.}
\usage{
score_clusters(km_clusters, GT_dir)
}
\arguments{
\item{km_clusters}{The k-means clusters generated by generate_clusters}

\item{GT_dir}{The directory holding the ground truth .csv files}
}
\value{
The scores table containing the GECO scored clusters
}
\description{
This function takes the k-means clusters generated by generate_clusters and
computes the GECO scores for each gene within each cluster. This scores table
is meant to be used with the final function in the GECO process:
assess_quality.
}
